# quicvarint: add Reader and Writer interfaces