This PR adds two interfaces to the `quicvarint` package: `Reader` and `Writer`, that are `io.Reader` + `io.ByteReader`, and `io.Writer` + `io.ByteWriter`, respectively, along with constructors that pass through the argument unchanged if it implements the necessary interfaces.

The interfaces to `quicvarint.Read` and `Write` are changed to accept a `Reader` and `Writer`, respectively. This change didn’t break tests in any package in this repo, nor require any changes in any package that depends on `quicvarint`.

Why? This is a small bit of support work for #3226, to ease construction of a package that processes H3 streams and datagrams outside of the `http3` package.

This PR also removes the existing `http3.byteReaderImpl` type, replacing it with `quicvarint.Reader`.