@marten-seemann 2021-07-20
Thank you for this PR.
Do you have any benchmarks for this change? The thing that worries me the most are the calls to `Read` and `Write`, as these functions are called **a lot**. If I understand correctly, `NewReader` and `NewWriter` shouldn't allocate when passed a `bytes.Reader` or a `bytes.Writer`, respectively (which is how we commonly use it for parsing the header and the frames).
Given that the actual work done in `Read` and `Write` is quite trivial though, my fear is that the type assertions might show up in a benchmark.