@ydnar 2021-07-21
> > When I ran the existing benchmarks suite, it seemed on par. I can add benchmarks for quicvarint.Read and Write if that’d be helpful.
> 
> Our benchmark package is a few years old and frankly, it's not very useful any more. I was thinking of a microbenchmark just for the `Read` and `Write` function here.

Sounds good. Would you prefer a gomega/ginkgo or vanilla Go benchmark?

> > If the type assertion does negatively impact the benchmarks, what do you think about changing the method signatures to accept a quicvarint.Reader or Writer instead of an io.Reader?
> 
> I wouldn't be opposed to that. I assume it doesn't break any (quic-go) code, does it?

It didn’t, since every existing caller passes a `*bytes.Buffer` to `Write`, which implements `quicvarint.Writer` and `Reader`.