@ydnar 2021-07-21
@marten-seemann looks like you’re right. The type assertion exacts a material penalty in the `Read` call, almost doubling its execution time:

```shell
❯ go test -bench . ./quicvarint
Running Suite: QUIC Varint Suite
================================
Random Seed: 1626839707
Will run 40 of 40 specs

••••••••••••••••••••••••••••••
------------------------------
• [MEASUREMENT]
Benchmarks
  with type assertion to quicvarint.Reader
    reading from a bytes.Reader

    Ran 100 samples:
    read:
      Fastest Time: 0.000s
      Slowest Time: 0.000s
      Average Time: 0.000s ± 0.000s
    read one (ns):
      Smallest: 17.913
       Largest: 64.634
       Average: 21.401 ± 5.843
    read all (ns):
      Smallest: 5732.000
       Largest: 20683.000
       Average: 6848.420 ± 1869.883
------------------------------
• [MEASUREMENT]
Benchmarks
  old Read(io.ByteReader) without type assertion
    reading from a bytes.Reader

    Ran 100 samples:
    read:
      Fastest Time: 0.000s
      Slowest Time: 0.000s
      Average Time: 0.000s ± 0.000s
    read one (ns):
      Smallest: 10.466
       Largest: 20.159
       Average: 11.693 ± 1.975
    read all (ns):
      Smallest: 3349.000
       Largest: 6451.000
       Average: 3741.860 ± 631.840
------------------------------
• [MEASUREMENT]
Benchmarks
  new Read(Reader) without type assertion
    reading from a bytes.Reader

    Ran 100 samples:
    read:
      Fastest Time: 0.000s
      Slowest Time: 0.000s
      Average Time: 0.000s ± 0.000s
    read one (ns):
      Smallest: 10.637
       Largest: 63.631
       Average: 11.613 ± 5.237
    read all (ns):
      Smallest: 3404.000
       Largest: 20362.000
       Average: 3716.000 ± 1675.817
------------------------------
••••••
Ran 40 of 40 Specs in 0.010 seconds
SUCCESS! -- 39 Passed | 0 Failed | 0 Pending | 0 Skipped
PASS
ok  	github.com/lucas-clemente/quic-go/quicvarint	0.145s
```