@ydnar 2021-07-22
Unrelated: do you plan to remove the CircleCI and Travis checks?