@ydnar 2021-07-26
@marten-seemann I put up a PR (#3238) that builds on this PR.

Do you want to review or merge this one?