@ydnar 2021-08-05
> Can you please remove the benchmark file? Ginkgo has deprecated the benchmark functionality anyway.

Done. 

This was my first time using Ginkgo. I normally write vanilla Go tests. What drove the decision to use it?